New: HLCOLOR attribute for IupTree on Windows and Motif (background of selected nodes). The selection color now will not change when the control loses its focus on Windows.
git-svn-id: svn://svn.code.sf.net/p/iup/iup/trunk@3116 2f08bc20-984a-4df8-b195-9b5d5d477537

+++ iup/src/gtk/iupgtk_tree.c
@@ -1281,7 +1281,7 @@ static int gtkTreeSetMarkedNodesAttrib(Ihandle* ih, const char* value)
   if (ih->data->mark_mode==ITREE_MARK_SINGLE || !value)
     return 0;
 
-  count = strlen(value);
+  count = (int)strlen(value);
   if (count > ih->data->node_count)
     count = ih->data->node_count;
